Moving some things around so there are no warnings from the bundle plugin.
