+++ src/main/java/org/mongojack/InitializationRequiredForTransformation.java
@@ -1,4 +1,4 @@
-package org.mongojack.internal.util;
+package org.mongojack;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
