+++ src/main/java/org/mongojack/JacksonCollectionKey.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mongojack.internal;
+package org.mongojack;
 
 import java.util.Objects;
