+++ src/main/java/org/mongojack/MongoCollectionDecorator.java
@@ -1,4 +1,4 @@
-package org.mongojack.internal.util;
+package org.mongojack;
 
 import com.mongodb.MongoNamespace;
 import com.mongodb.ReadConcern;
