+++ src/main/java/org/mongojack/QueryCondition.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.mongojack.internal.query;
+package org.mongojack;
 
 public interface QueryCondition {
 }
