+++ src/main/java/org/mongojack/internal/update/MultiUpdateOperationValue.java
@@ -16,6 +16,8 @@
  */
 package org.mongojack.internal.update;
 
+import org.mongojack.UpdateOperationValue;
+
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
